requirements.txt
appdirs==1.4.4
asttokens==2.1.0
attrs==22.1.0
audioread==3.0.0
backcall==0.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
ffmpeg-python==0.2.0
fonttools==4.38.0
future==0.18.2
idna==3.4
ipykernel==6.17.1
ipython==8.6.0
jedi==0.18.1
joblib==1.2.0
jsonschema==4.17.0
jupyter_client==7.4.5
jupyter_core==5.0.0
kiwisolver==1.4.4
librosa==0.9.2
llvmlite==0.39.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
musdb==0.4.0
museval==0.4.0
nest-asyncio==1.5.6
numba==0.56.4
numpy==1.23.4
packaging==21.3
pandas==1.5.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.3
pooch==1.6.0
prompt-toolkit==3.0.32
psutil==5.9.4
pure-eval==0.2.2
pyaml==21.10.1
pycparser==2.21
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.6
pywin32==305
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
resampy==0.4.2
scikit-learn==1.1.3
scipy==1.9.3
simplejson==3.17.6
six==1.16.0
soundfile==0.11.0
stack-data==0.6.0
stempeg==0.2.3
threadpoolctl==3.1.0
torch==1.13.0
torchaudio==0.13.0
torchlibrosa==0.0.9
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
